+++ src/libretro/config.hpp
@@ -60,32 +60,6 @@ namespace melonds {
         Arm9i,
     };
 
-    constexpr std::string_view ConsoleTypeName(ConsoleType type) noexcept {
-        switch (type) {
-            case ConsoleType::DS:
-                return "DS";
-            case ConsoleType::DSi:
-                return "DSi";
-        }
-    }
-
-    constexpr std::string_view ConsoleTypeName(SPI_Firmware::FirmwareConsoleType type) noexcept {
-        switch (type) {
-            case SPI_Firmware::FirmwareConsoleType::DS:
-                return "DS";
-            case SPI_Firmware::FirmwareConsoleType::DSi:
-                return "DSi";
-            case SPI_Firmware::FirmwareConsoleType::DSLite:
-                return "DS Lite";
-            case SPI_Firmware::FirmwareConsoleType::iQueDSLite:
-                return "iQue DS Lite";
-            case SPI_Firmware::FirmwareConsoleType::iQueDS:
-                return "iQue DS";
-            default:
-                return "unknown";
-        }
-    }
-
     enum class ScreenSwapMode {
         Hold,
         Toggle,